fix CI build
